refactor(ai-chat): reorganize module structure (#50)
