+++ app/ai-chat/src/platform.rs
@@ -1,0 +1,3 @@
+pub(crate) mod capture;
+pub(crate) mod display;
+pub(crate) mod gpui_ext;
