+++ app/ai-chat/src/platform/capture.rs
