+++ app/ai-chat/src/platform/display.rs
