+++ app/ai-chat/src/platform/gpui_ext.rs
