+++ app/ai-chat/src/state/chat.rs
