+++ app/ai-chat/src/state/chat/models.rs
