+++ app/ai-chat/src/state/chat/tree.rs
