+++ app/ai-chat/src/views/conversation.rs
@@ -1,0 +1,2 @@
+pub(crate) mod detail;
+pub(crate) mod preview;
