+++ app/ai-chat/src/views/conversation/detail.rs
