+++ app/ai-chat/src/views/conversation/preview.rs
