+++ app/ai-chat/src/views/screenshot.rs
@@ -1,0 +1,7 @@
+use gpui::App;
+
+pub(crate) mod overlay;
+
+pub(crate) fn init(cx: &mut App) {
+    overlay::init(cx);
+}
